脚手架提供@Slf4j和@Log*装饰器的集成和工具方法
